Render the model's unsigned link values through FormatUint in networkd
Replace strconv.Itoa with a decimal helper over uint64 at every site in
internal/networkd/networkd.go that renders the VLAN id, the prefix
length, a route metric, or the router lifetime, and type the test
fixtures' metrics and lifetime as uint32 in internal/networkd.

+++ internal/networkd/spec_test.go
@@ -167,7 +167,7 @@ func TestValidateOccupiesTheLeaseMetricOnlyWithoutAGateway(t *testing.T) {
 
 			static := staticLinkSpec(networkSection(tc.section, tc.key, "100"))
 			static.IPv6.Gateway = netip.MustParseAddr("2001:db8::1")
-			static.IPv6.RouteMetric = new(10)
+			static.IPv6.RouteMetric = new(uint32(10))
 			if err := networkd.Validate(static); err != nil {
 				t.Fatalf("Validate rejected a lease metric on a family with a gateway: %v", err)
 			}
